--- src/main/java/com/lubycon/curriculum/curriculum/repository/BlogLinkRepository.java
@@ -1,8 +1,0 @@
-package com.lubycon.curriculum.curriculum.repository;
-
-import com.lubycon.curriculum.curriculum.domain.BlogLink;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface BlogLinkRepository extends JpaRepository<BlogLink, Long> {
-
-}
