+++ src/test/java/com/lubycon/curriculum/blog/BlogRepositoryTest.java
@@ -1,20 +1,21 @@
-package com.lubycon.curriculum.curriculum.repository;
+package com.lubycon.curriculum.blog;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.lubycon.curriculum.base.RepositoryTest;
-import com.lubycon.curriculum.curriculum.domain.BlogLink;
+import com.lubycon.curriculum.blog.domain.Blog;
+import com.lubycon.curriculum.blog.repository.BlogRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
-class BlogLinkRepositoryTest extends RepositoryTest {
+class BlogRepositoryTest extends RepositoryTest {
 
   @Autowired
-  private BlogLinkRepository blogLinkRepository;
+  private BlogRepository blogRepository;
 
-  private BlogLink blogLink;
+  private Blog blog;
 
   private static final String TITLE = "JUnit5의 기본적인 사용법들";
   private static final String LINK = "https://shinsunyoung.tistory.com/100";
@@ -22,7 +23,7 @@ class BlogLinkRepositoryTest extends RepositoryTest {
   @BeforeEach
   public void setUp() {
 
-    blogLink = blogLinkRepository.save(BlogLink.builder()
+    blog = blogRepository.save(Blog.builder()
         .title(TITLE)
         .link(LINK)
         .build());
@@ -31,7 +32,7 @@ class BlogLinkRepositoryTest extends RepositoryTest {
   @DisplayName("findById() 테스트")
   @Test
   public void findById() {
-    BlogLink link = blogLinkRepository.findById(this.blogLink.getId()).get();
+    Blog link = blogRepository.findById(this.blog.getId()).get();
 
     assertThat(link.getTitle()).isEqualTo(TITLE);
     assertThat(link.getLink()).isEqualTo(LINK);
